Implementacao de desconto

+++ geekShopping/GeekShopping/GeekShopping.CouponAPI/Data/ValueObjects/CouponVO.cs
@@ -2,6 +2,6 @@
     public class CouponVO {
         public long Id { get; set; }
         public string CouponCode { get; set; }
-        public decimal DiscountAmount { get; set; }
+        public double DiscountAmount { get; set; }
     }
 }
